requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
gunicorn==21.2.0
Werkzeug==3.0.1
pyngrok==7.1.0
